Moved user view model method into orchestrator folder, replaced newing up orchestrators in the controller with DI
