+++ src/SFA.DAS.AdminService.Web/Orchestrators/UserViewModelOrchestrator.cs
@@ -5,15 +5,15 @@ using System;
 using System.Threading.Tasks;
 using static StackExchange.Redis.Role;
 
-namespace SFA.DAS.AdminService.Web.Helpers
+namespace SFA.DAS.AdminService.Web.Orchestrators
 {
-    public class UserViewModelHelper : IUserViewModelHelper
+    public class UserViewModelOrchestrator : IUserViewModelOrchestrator
     {
         private readonly IOrganisationsApiClient _organisationsApiClient;
         private readonly IContactsApiClient _contactsApiClient;
         private readonly IMapper _mapper;
 
-        public UserViewModelHelper(IContactsApiClient contactsApiClient, IOrganisationsApiClient organisationsApiClient, IMapper mapper)
+        public UserViewModelOrchestrator(IContactsApiClient contactsApiClient, IOrganisationsApiClient organisationsApiClient, IMapper mapper)
         {
             _contactsApiClient = contactsApiClient;
             _organisationsApiClient = organisationsApiClient;
